% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_SwapParRates.R
\docType{data}
\name{DB_SwapParRates}
\alias{DB_SwapParRates}
\title{Swap Par Rates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2501 rows and 7 columns.
}
\usage{
DB_SwapParRates
}
\description{
Dataset used in A. Meucci (2008) "Estimation of Structured T-Copulas".
}
\keyword{datasets}
