% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Log-Likelihood for t-distribution}
\usage{
log_lik(x, nu, sigma)
}
\arguments{
\item{x}{A numeric matrix.}

\item{nu}{A numeric scalar.}

\item{sigma}{A dispersion matrix.}
}
\value{
A \code{double} with the estimated likelihood.
}
\description{
Computes the Maximum Likelihood function for multivariate
t-distributed scenarios.
}
\keyword{internal}
