% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLik.R
\name{LogLik}
\alias{LogLik}
\title{Log-Likelihood for t-distribution}
\usage{
LogLik(x, Nu, Sigma)
}
\arguments{
\item{x}{A numeric matrix.}

\item{Nu}{A numeric scalar.}

\item{Sigma}{A correlation matrix.}
}
\value{
A single scalar.
}
\description{
This function computes the Maximum Likelihood function for multivariate
t-distributed scenarios.
}
\examples{
#
}
\keyword{internal}
