% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_recursion_for_t.R
\name{mle_recursion_for_t}
\alias{mle_recursion_for_t}
\title{Computes the Correlation ML Estimator}
\usage{
mle_recursion_for_t(x, nu, k, tolerance = 1e-10)
}
\arguments{
\item{x}{A numeric matrix.}

\item{nu}{A numeric scalar.}

\item{k}{A numeric scalar.}

\item{tolerance}{A numeric scalar.}
}
\value{
A correlation matrix.
}
\description{
This function recursively computes the ML estimators of a correlation matrix
with t-copula and isotropic structure.
}
\keyword{internal}
