% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MleRecursionForT.R
\name{MleRecursionForT}
\alias{MleRecursionForT}
\title{Computes ML Estimator of the Correlation Matrix}
\usage{
MleRecursionForT(x, Nu, K, Tolerance = 10^(-10))
}
\arguments{
\item{x}{A numeric matrix.}

\item{Nu}{A numeric scalar.}

\item{K}{A numeric scalar.}

\item{Tolerance}{A numeric scalar.}
}
\value{
A correlation matrix.
}
\description{
This function recursively computes the ML estimators of the correlation matrix
of a t-copula with isotropic structure.
}
\examples{
#
}
\keyword{internal}
