% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_copula_marginal.R
\name{separate_copula_marginal}
\alias{separate_copula_marginal}
\title{Separates the Copula from the Marginal Distribution}
\usage{
separate_copula_marginal(X)
}
\arguments{
\item{X}{A numeric \code{matrix}.}
}
\value{
A \code{list} with 3 components:
    \itemize{
      \item \code{W}: the ordered marginals
      \item \code{F}: the CDF
      \item \code{U}: the copula
    }
}
\description{
This function decomposes the the pure "joint" from the pure "marginal"
features of a multivariate distribution.
}
\keyword{internal}
