% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateMargCop.R
\name{SeparateMargCop}
\alias{SeparateMargCop}
\title{Separates the Copula from the Marginal Distribution}
\usage{
SeparateMargCop(X)
}
\arguments{
\item{X}{A numeric \code{matrix}.}
}
\value{
A \code{list} with 3 components:
}
\description{
This function decomposes the the pure "joint" from the pure "marginal"
features of a multivariate distribution.
}
\examples{
#
}
\keyword{internal}
