% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_t_mle.R
\name{struct_t_mle}
\alias{struct_t_mle}
\title{ML Estimators for a Structured Correlation Matrix}
\usage{
struct_t_mle(x, k, tolerance = 1e-10)
}
\arguments{
\item{x}{A \code{matrix} with multivariate i.i.d. variables.}

\item{k}{A single \code{double} with the number of relevant factors.}

\item{tolerance}{A single \code{double} used for convergence.}
}
\value{
A \code{list} with 2 components:
    \itemize{
      \item \code{Nu}: the degrees of freedom
      \item \code{C}: the correlation matrix
    }
}
\description{
Computes the the correlation matrix and the degrees of freedom of a t-copula
using the Maximum-Likelihood (ML) estimator.
}
\examples{
data("swap")

# The first difference is stationary
swap_diff <- swap[2:nrow(swap), ] - swap[1:nrow(swap) - 1, ]

# fit the model
struct_t_mle(x = swap_diff, k = 3)
}
