% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrucTMLE.R
\name{StrucTMLE}
\alias{StrucTMLE}
\title{ML Estimators for a Structured Correlation Matrix}
\usage{
StrucTMLE(X, K, Tolerance = 10^(-10))
}
\arguments{
\item{X}{A numeric matrix.}

\item{K}{A numeric scalar.}

\item{Tolerance}{A numeric scalar.}
}
\value{
A \code{list} with 2 components: Nu (the degrees of freedom) and C (the
correlation matrix).
}
\description{
This function recursively computes the ML estimators of the correlation matrix
and the degrees of freedom of a t-copula with isotropic structure.
}
\examples{
#
}
\references{
Meucci, Attilio, Estimation of Structured T-Copulas (April 2008). Available
at SSRN: \url{https://ssrn.com/abstract=1126401} or \url{http://dx.doi.org/10.2139/ssrn.1126401}.

Attilio Meucci (2020). \href{Estimation of Structured t-Copulas}{https://www.mathworks.com/matlabcentral/fileexchange/19751-estimation-of-structured-t-copulas}, MATLAB Central File Exchange. Retrieved October 14, 2020.
}
