% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap.R
\docType{data}
\name{swap}
\alias{swap}
\title{Swap Par Rates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2501 rows and 7 columns.
}
\usage{
swap
}
\description{
Dataset used in A. Meucci (2008): "Estimation of Structured T-Copulas".
}
\keyword{datasets}
