% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcopula.R
\docType{package}
\name{tcopula}
\alias{tcopula}
\title{tcopula: A package for computing structured t-copulas}
\description{
The package implements the algorithm described in "Estimation of Structured
t-Copulas".
}
\references{
Meucci, Attilio, Estimation of Structured T-Copulas (April 2008). Available
at SSRN: \url{https://ssrn.com/abstract=1126401} or \url{http://dx.doi.org/10.2139/ssrn.1126401}.

Attilio Meucci (2020). \href{Estimation of Structured t-Copulas}{https://www.mathworks.com/matlabcentral/fileexchange/19751-estimation-of-structured-t-copulas}, MATLAB Central File Exchange. Retrieved October 14, 2020.
}
