% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrucTMLE.R
\name{StrucTMLE}
\alias{StrucTMLE}
\title{ML Estimators for a Structured Correlation Matrix}
\usage{
StrucTMLE(X, K, Tolerance = 10^(-10))
}
\arguments{
\item{X}{A \code{matrix} made of timeseries invariants.}

\item{K}{A \code{numeric} scalar with the number of factors.}

\item{Tolerance}{A \code{numeric} scalar used for convergence. By default,
the algorithm stops it's computations when the error is smaller than \code{10^(-10)}.}
}
\value{
A \code{list} with 2 components: \code{Nu} (the degrees of freedom)
and \code{C} (the correlation matrix).
}
\description{
Recursively computes the the correlation matrix and the degrees of freedom
by Maximum-Likelihood (ML) of a t-copula with isotropic structure.
}
\examples{
data("DB_SwapParRates")
Rates <- DB_SwapParRates

# The first difference is stationary
X <- Rates[2:nrow(Rates), ] - Rates[1:nrow(Rates) - 1, ]

# fit the model
StrucTMLE(X = X, K = 3)
}
\references{
Meucci, Attilio, Estimation of Structured T-Copulas (April 2008). Available
at SSRN: \url{https://ssrn.com/abstract=1126401} or \url{http://dx.doi.org/10.2139/ssrn.1126401}.

Attilio Meucci (2020). \href{Estimation of Structured t-Copulas}{https://www.mathworks.com/matlabcentral/fileexchange/19751-estimation-of-structured-t-copulas}, MATLAB Central File Exchange. Retrieved October 14, 2020.
}
